requirements.txt
Django==5.0.4
django-tenants==3.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.3
django-redis==5.4.0
stripe==9.3.0
WeasyPrint==62.0
django-filter==24.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.3.0
django-crispy-forms==2.1
crispy-tailwind==1.0.3
django-htmx==1.18.0
whitenoise==6.6.0
sentry-sdk==2.1.1
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.5.0

# Runtime/supporting deps (required by Docker/CI instructions)
gunicorn==22.0.0
django-celery-beat==2.6.0

# Dev tooling (required by CI instructions)
black==24.4.2
flake8==7.0.0
